test(postgres): fold ad-hoc integration tests into e2e.rs
The branch's engine_crud and row_ref integration tests duplicated
harness code (test_url fallback, no PRAX_E2E gate, no unique
table names) that the rebased develop now provides centrally via
tests/e2e.rs. Merge both into e2e.rs as e2e_query_many_typed_
decodes_rows and e2e_row_ref_primitive_reads, delete the standalone
files, and adopt PRAX_E2E / POSTGRES_URL / unique_table / ignore
conventions so the compose test-postgres runner picks them up.
